requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.0
async-timeout==5.0.1
asyncio==4.0.0
attrs==25.4.0
backports-datetime-fromisoformat==2.0.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.1.0
dj-email-url==1.0.6
Django==5.2.9
django-allauth==65.13.1
django-cache-url==3.4.5
environs==14.5.0
frozenlist==1.8.0
idna==3.11
isodate==0.7.2
lxml==6.0.2
marshmallow==4.2.0
melipayamak==1.0.1
multidict==6.7.0
pillow==12.1.0
platformdirs==4.5.1
propcache==0.4.1
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
yarl==1.22.0
zeep==4.3.2
psycopg2-binary
gunicorn
aiopg
psutil
apscheduler
aiosqlite
